looker: Update heap size calculation for cgroups
When running within Docker, it's less than ideal to use /proc/meminfo since this is the memory on the entire node. Not what has been allocated to the running container. For that, we need to yank the data out of the cgroup limit instead.

For inspiration, this was mostly yanked out of what the RabbitMQ docker container does: https://github.com/docker-library/rabbitmq/blob/master/3.7/alpine/docker-entrypoint.sh#L300-L316